Added Lamia species
Static methods can be used to create species with common traits as
constuctor arguments. They can then simply be called as the species
name. It is abstract enough that any species can be added this way, an
example has been given, with a new trait added to accomodate.
Instructions included.