Added "ref" variable to reference images
Conitnue to use the "Name Mood" convention to name all character images
in the future. Whenever interaction is called, the name of the character
for finding images must be passed to the 'reference' argument so that
the image can be called though standard string manipulation using the
variable 'ref' for the current character.